MIM-10 拆分 guidance fallback 回归测试
